Activate button is now controlled with permission PRJ_ACTIVATE_USER
git-svn-id: https://square.dwsbox.com/svn/dws/cpweb/trunk@3343 e88eb70a-f57a-4e2b-9716-5226b8acd346

+++ src/main/java/se/dabox/cocobox/cpweb/module/project/ProjectRosterJsonGenerator.java
@@ -113,7 +113,7 @@ class ProjectRosterJsonGenerator {
                     g.writeStringField("displayName", name);
                     g.writeStringField("email", ua.getPrimaryEmail());
                 }
-                boolean main = mainUserId != null && mainUserId.longValue() == item.getUserId();
+                boolean main = mainUserId != null && mainUserId == item.getUserId();
                 g.writeBooleanField("main", main);
                 writeDateField("firstEmail", item.getFirstMail());
                 writeDateField("lastEmail", item.getLastMail());
